pyproject.toml
[tool.poetry]
name = "newsbot"
version = "0.1.0"
description = ""
authors = ["Dmitry Mogilevsky <dmitry.mogilevsky@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = ">=0.115.13,<0.116.0"
uvicorn = ">=0.34.3,<0.35.0"
redis = ">=6.2.0,<7.0.0"
pyyaml = ">=6.0.2,<7.0.0"
googlenews = ">=1.6.15,<2.0.0"
googlenewsdecoder = ">=0.1.7,<0.2.0"
langdetect = ">=1.0.9,<2.0.0"
newspaper3k = ">=0.2.8,<0.3.0"
beautifulsoup4 = ">=4.12.0,<5.0.0"
trafilatura = ">=1.12.0,<2.0.0"
requests = ">=2.32.0,<3.0.0"
openai = ">=1.88.0,<2.0.0"
blueskysocial = ">=1.9.0,<2.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.0"
mypy = "^1.16.1"
pre-commit = "^4.2.0"
pytest = "^8.4.1"
types-pyyaml = "^6.0.12.20250516"
isort = "^1.1.0"
types-requests = "^2.32.4.20250611"

[tool.pylint.master]
ignore-patterns = ["^.*\\.pyi$"]

[tool.pylint.messages_control]
disable = ["W0718","missing-module-docstring","line-too-long","too-many-locals","too-few-public-methods","fixme"]